move GCSSourceFactory
